+++ src/users/tasks.py
@@ -1,11 +1,12 @@
 from celery import shared_task
+from django.http import HttpRequest
 
-from users.services.emails import send_registration_email, send_reset_password
+from users.services.emails import send_activation_email, send_reset_password
 
 
 @shared_task
-def verify_email(user_instance):
-    send_registration_email(user_instance)
+def verify_email(request: HttpRequest, user_id: int) -> None:
+    send_activation_email(request, user_id)
 
 
 @shared_task